+++ legadilo/feeds/views/feed_articles_view.py
@@ -49,4 +49,5 @@ def feed_articles_view(
         request,
         Feed.objects.get_articles(feed),
         _("Articles of feed '%(feed_title)s'") % {"feed_title": feed.title},
+        extra_ctx={"feed": feed},
     )