+++ legadilo/reading/models/reading_list.py
@@ -250,8 +250,7 @@ class ReadingList(models.Model):
         )
 
     def save(self, *args, **kwargs):
-        if not self.slug:
-            self.slug = slugify(self.title)
+        self.slug = slugify(self.title)
         return super().save(*args, **kwargs)
 
     def delete(self, *args, **kwargs):